Rename parseOperatorType to parsePunctuator
