+++ frontend/src/metabase-lib/v1/expressions/tokenizer/index.ts
@@ -1,2 +1,1 @@
 export * from "./tokenize";
-export * from "./types";
