--- frontend/src/metabase-lib/v1/expressions/tokenizer/types.ts
@@ -1,18 +1,0 @@
-export enum PUNCTUATOR {
-  Comma = ",",
-  OpenParenthesis = "(",
-  CloseParenthesis = ")",
-  Plus = "+",
-  Minus = "-",
-  Multiply = "*",
-  Divide = "/",
-  Equal = "=",
-  NotEqual = "!=",
-  LessThan = "<",
-  GreaterThan = ">",
-  LessThanEqual = "<=",
-  GreaterThanEqual = ">=",
-  Not = "not",
-  And = "and",
-  Or = "or",
-}
